Moved actions and reducers and cleaned up
