+++ src/app/store/auth/actions.js
@@ -60,3 +60,32 @@ export const saveLoggedUserSession = user => dispatch => {
 
 export const updateInput = (field, value) => dispatch =>
   dispatch({ type: UPDATE_INPUT, payload: { field, value } });
+
+export const registerNewAccount = props => async dispatch => {
+  const { email, password, userName, wages } = props;
+  try {
+    const user = await firebase
+      .auth()
+      .createUserWithEmailAndPassword(email, password);
+
+    const currentUser = firebase.auth().currentUser;
+
+    currentUser.sendEmailVerification();
+
+    const { uid } = user.user;
+
+    await firebase
+      .firestore()
+      .collection("users")
+      .doc(uid)
+      .set({
+        uid,
+        userName,
+        wages
+      });
+
+    dispatch({ type: "REGISTRATION COMPLETE" });
+  } catch (error) {
+    console.error(error);
+  }
+};
