+++ src/app/store/reducers.js
@@ -1,15 +1,15 @@
 import { combineReducers } from "redux";
 
-import today from "../modules/DailyForm/reducer";
-import auth from "../modules/LoginPage/reducer";
-import sideDrawer from "../components/SideDrawer/reducer";
-import reports from "../modules/ReportScreen/reducer";
-import settings from "../modules/SettingsScreen/reducer";
+import today from "./today/reducer";
+import auth from "./auth/reducer";
+import ui from "./ui/reducer";
+import reports from "./reports/reducer";
+import settings from "./settings/reducer";
 
 export default combineReducers({
   today,
   auth,
-  sideDrawer,
+  ui,
   reports,
   settings
 });
