+++ src/app/store/ui/reducer.js
@@ -1,0 +1,7 @@
+import { combineReducers } from "redux";
+
+import sideDrawer from "./sideDrawer/reducer";
+
+export default combineReducers({
+  sideDrawer
+});
