+++ src/app/store/ui/sideDrawer/reducer.js
